add: Basic function implementation complete #33
